비즈니스와 관련없는 hook 별도 폴더로 이동
